fix: changing ordersQuantity to Long primitive type

+++ src/main/java/me/gfumagali/btgpedidos/model/documents/ClientOrders.java
@@ -16,7 +16,7 @@ public class ClientOrders {
     @Id
     private Long clientCode;
 
-    private Integer ordersQuantity;
+    private Long ordersQuantity;
 
     private List<Order> orders;
 }
